refactor(live-editor): split into multiple files
